refactor(frontend): extract AuthState and helpers to dedicated file
Move AuthState class, isLeadershipRole(), and getCurrencySymbol() out
of auth.dart into auth_state.dart for testability without pulling in
the full Auth0 dependency tree.
